requirements.txt
streamlit
langchain
google-generativeai
chromadb
pypdf
langchain-google-genai
langchain-community
python-dotenv